+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_de.dart
@@ -138,6 +138,12 @@ class StreamChatLocalizationsDe extends GlobalStreamChatLocalizations {
       'Die Dateigröße ist begrenzt auf $limitInMB MB.';
 
   @override
+  String fileTypeNotSupportedError(String? extension) {
+    if (extension != null) return "'.$extension'-Dateien werden nicht zum Hochladen unterstützt.";
+    return 'Dieser Dateityp wird nicht zum Hochladen unterstützt.';
+  }
+
+  @override
   String get addAFileLabel => 'Datei hinzufügen';
 
   @override